[feature] session_state CLI 응집 하위 모듈 재분할 + 역방향 헬퍼 core 복귀
## 관련 이슈 번호
#929

## 작업내용
- session_state_cli.py(2190줄) 를 응집 단위로 재분할 → cli.py 1349줄:
  - session_state_cli_wave.py 신규 (331줄) — wave board + merge lock (peer 병렬)
  - session_state_cli_finalize.py 신규 (596줄) — step 완료 + run finalization
- 순수 step-state 헬퍼(_read_steps_jsonl / _steps_jsonl_path) 를 core 로 복귀 →
  facade→cli 역방향 위임 래퍼 제거, 의존 단방향화 (cli → core)
- cli.py 가 이동 핸들러를 re-expose → facade 재수출/CLI 계약 무변경
- 분할 모듈 순환 import 회귀 테스트에 신규 2개 모듈 추가

## Test Plan
- [x] 전체 unittest 1631 OK (분할 전후 동일 카운트)
- [x] guard eval 33/33, cross-ref PASS, static-quality(ruff/mypy/bandit) PASS
- [x] 이동 함수 27개 byte-identical (순수 이동 확증)
